+++ src/app/core/services/gen-id.generator.ts
@@ -1,0 +1,5 @@
+export const genId = (function (): () => number {
+  let id = 0;
+
+  return (): number => (id += 1);
+})();